requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
click==8.1.8
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
greenlet==3.2.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mysqlclient==2.2.7
numpy==2.2.6
openai==1.82.0
pandas==2.2.3
pip==24.0
priority==2.0.0
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic-core==2.33.2
pygments==2.19.1
pymysql==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
rich==14.0.0
rich-toolkit==0.14.6
scikit-learn==1.6.1
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
threadpoolctl==3.6.0
tqdm==4.67.1
typer==0.15.4
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wsproto==1.2.0
